feat(customer): migrate Kunde list and edit pages to projectforge-next
Hand-built customer (Kunde) list and edit in projectforge-next, layout-free
on the backend (CustomerPagesRest now extends AbstractDTOEntityRest, no
createListLayout/createEditLayout). status is a pinned default magic filter;
the duplicate-number check moved to validate().

- Customer.nummer changed Int? -> Long? to match KundeDO.nummer: BaseDTO.copy
  only transfers fields of equal type, so the id was silently dropped, causing
  IdentifierGenerationException when saving a new customer.
- Unmount the Wicket customer pages (WebRegistry) and drop the menu registration
  (MenuItemRegistry); the menu resolves to next/customer, wa/customer* is
  redirected by OrphanedLinkFilter. CustomerListPage/EditPage classes are kept
  for the select panels used by the still-Wicket Projekt/UserPref edit forms.
- Register customer under NextMigration.MIGRATED and HAND_BUILT_CATEGORIES;
  add WRITE:customer to the FINANCE_WRITE 2FA shortcut.
- Customer favorites are deliberately not carried over.
- e2e spec plus a seeded customer fixture.

+++ projectforge-rest/src/main/kotlin/org/projectforge/rest/dto/Customer.kt
@@ -30,7 +30,7 @@ import org.projectforge.framework.i18n.translate
 
 class Customer(id: Long? = null,
                displayName: String? = null,
-               var nummer: Int? = null,
+               var nummer: Long? = null,
                var name: String? = null,
                var identifier: String? = null,
                var division: String? = null,
